RW-571: Added grid cluster targeted execution support.
Grid node "cluster" names, used to denote the runtime characteristics of a node, can be assigned by the DeployR admin to individual nodes or groups of nodes on the DeployR grid, for example, "hi-mem" or "hi-cpu".

This update to the client library allows client applications request tasks be executed on nodes within a specific cluster on the grid:

- RUser.createProject
- RUser.createProjectPool
- RUser.submitJob*
- RClient.executeScript
- RClient.executeExternal
- RClient.renderScript

This feature is DeployR Enterprise only.