RW-754: Added self-signed SSL cert support.
The DeployR server supports and can be configured to require HTTPS connections from clients. Sometimes (primarily for testing or prototyping) self-signed SSL certificates are used on the server. The client library has been updated to work with servers using self-signed certs:

- RClientFactory.createClient(endpoint, allowSelfSignedSSLCert )

To facilitate these changes instances of URL are no longer returned on download calls on the API. In place of URLs, instances of InputStream are returned on the following interfaces:

- InputStream RProject.export()
- InputStream RProjectFile.download()
- InputStream RProjectResult.download()
- InputStream RRepositoryFile.download()

The underlying URL is still avaiable on the *Details instance associated with the File or Result.

All unit tests and tutorial examples have been updated to reflect these changes.