RW-777: Added RDataTable support to RDataFactory.
The new RDataTable adds support that grealy simplifies working with tabular data in client applications.

An RDataTable can ingest the data within a delimiter-separated file and expose that data as tabular data. It can also ingest RDataFrame, R*Matrix and R*Vector and expose the data found within those encoded R objects as tabular data.

An RDataTable can also emit it's tabular data as an RDataFrame, R*Matrix or an R*Vector. The RData emitted by an RDataTable can then be passed as a DeployR-encoded R object on API calls.